+++ agential/cog/functional/reflexion.py
@@ -9,7 +9,7 @@ from langchain_core.messages.human import HumanMessage
 from langchain_core.prompts.prompt import PromptTemplate
 from tiktoken.core import Encoding
 
-from agential.cog.prompts.agents.reflexion import (
+from agential.cog.prompts.agent.reflexion import (
     LAST_TRIAL_HEADER,
     REFLECTION_HEADER,
     REFLEXION_COT_INSTRUCTION,
