+++ agential/cog/modules/reflect/reflexion.py
@@ -11,7 +11,7 @@ from agential.cog.functional.reflexion import (
     react_reflect,
 )
 from agential.cog.modules.reflect.base import BaseReflector
-from agential.cog.prompts.agents.reflexion import (
+from agential.cog.prompts.agent.reflexion import (
     REFLECTION_AFTER_LAST_TRIAL_HEADER,
     REFLEXION_COT_REFLECT_INSTRUCTION,
     REFLEXION_COT_REFLECT_INSTRUCTION_NO_CONTEXT,
