+++ tests/cog/agent/test_self_refine.py
@@ -4,7 +4,7 @@ from langchain_community.chat_models.fake import FakeListChatModel
 from langchain_core.language_models.chat_models import BaseChatModel
 
 from agential.cog.agent.self_refine import SelfRefineAgent
-from agential.cog.prompts.agents.self_refine import (
+from agential.cog.prompts.agent.self_refine import (
     GSM8K_CRITIQUE_FEWSHOT_EXAMPLES,
     GSM8K_REFINE_FEWSHOT_EXAMPLES,
     SELF_REFINE_CRITIQUE_INSTRUCTION_GSM8K,
