+++ tests/cog/strategies/react/test_code.py
@@ -4,7 +4,7 @@ from langchain_community.chat_models.fake import FakeListChatModel
 from langchain_core.language_models.chat_models import BaseChatModel
 from tiktoken import Encoding
 
-from agential.cog.prompts.agents.react import REACT_INSTRUCTION_MBPP
+from agential.cog.prompts.agent.react import REACT_INSTRUCTION_MBPP
 from agential.cog.prompts.benchmark.mbpp import MBPP_FEWSHOT_EXAMPLES_REACT
 from agential.cog.strategies.react.code import (
     ReActCodeStrategy,
