+++ tests/cog/strategies/self_refine/test_math.py
@@ -2,7 +2,7 @@
 
 from langchain_community.chat_models.fake import FakeListChatModel
 
-from agential.cog.prompts.agents.self_refine import (
+from agential.cog.prompts.agent.self_refine import (
     GSM8K_CRITIQUE_FEWSHOT_EXAMPLES,
     GSM8K_REFINE_FEWSHOT_EXAMPLES,
     SELF_REFINE_CRITIQUE_INSTRUCTION_GSM8K,
